feat: sanity integration (#502)
* ssg quries

* resolved build errors
