+++ app/(withlayout)/annual_reports/page.tsx
@@ -1,14 +1,29 @@
 import { Metadata } from 'next';
 import Reports from './Reports';
 
-export default function Page() {
+
+import { client } from '@/lib/sanity/client';
+import { annualReportsStructure } from '@/data/annual_reports';
+
+const queryReport = '*[_type == "annualReport"]';
+
+async function getData() {
+  const data = (await client.fetch(queryReport)) as any[];
+  console.log('data:'+data.sort((a, b) => b.id - a.id));
+  return data;
+}
+
+export default async function Page() {
+
+  const data = await getData();
+
   return (
     <div className="w-full flex justify-center">
       <div className="w-fit max-w-5xl p-4 flex flex-col gap-4 pb-12">
         <p className="text-dwd-primary font-bold text-4xl text-center mb-2 mt-2">
           Annual Reports
         </p>
-        <Reports />
+        <Reports Fulldata={data} />
       </div>
     </div>
   );
