+++ components/campus/MainContent.tsx
