+++ data/krc/dataTelFull.ts
